--- src/main/java/blackJack/Application.java
@@ -1,11 +1,0 @@
-package blackJack;
-
-import blackJack.controller.BlackJackController;
-
-public class Application {
-
-    public static void main(String[] args) {
-        final BlackJackController blackJackController = new BlackJackController();
-        blackJackController.run();
-    }
-}
