+++ src/main/java/blackjack/Application.java
@@ -1,0 +1,11 @@
+package blackjack;
+
+import blackjack.controller.BlackjackController;
+
+public class Application {
+
+    public static void main(String[] args) {
+        final BlackjackController blackjackController = new BlackjackController();
+        blackjackController.run();
+    }
+}
