--- src/test/java/blackJack/domain/card/DeckTest.java
@@ -1,15 +1,0 @@
-package blackJack.domain.card;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class DeckTest {
-
-    @Test
-    @DisplayName("카드덱 생성 테스트")
-    void createValidDeck() {
-        assertThat(Deck.create()).isNotNull();
-    }
-}
