+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,0 +1,15 @@
+package blackjack.domain.card;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class DeckTest {
+
+    @Test
+    @DisplayName("카드덱 생성 테스트")
+    void createValidDeck() {
+        assertThat(Deck.create()).isNotNull();
+    }
+}
